+++ src/titiler/core/titiler/core/factory.py
@@ -37,7 +37,7 @@ from rio_tiler.types import ColorMapType
 from rio_tiler.utils import CRS_to_uri
 from starlette.requests import Request
 from starlette.responses import HTMLResponse, Response
-from starlette.routing import Match, compile_path, replace_params
+from starlette.routing import Match, NoMatchFound, compile_path, replace_params
 from starlette.templating import Jinja2Templates
 from typing_extensions import Annotated
 
@@ -70,7 +70,7 @@ from titiler.core.dependencies import (
     TileParams,
 )
 from titiler.core.models.mapbox import TileJSON
-from titiler.core.models.OGC import TileMatrixSetList
+from titiler.core.models.OGC import TileMatrixSetList, TileSet, TileSetList
 from titiler.core.models.responses import (
     ColorMapsList,
     InfoGeoJSON,
@@ -130,6 +130,7 @@ class BaseFactory(metaclass=abc.ABCMeta):
     Attributes:
         router (fastapi.APIRouter): Application router to register endpoints to.
         router_prefix (str): prefix where the router will be mounted in the application.
+        route_dependencies (list): Additional routes dependencies to add after routes creations.
 
     """
 
@@ -224,19 +225,23 @@ class TilerFactory(BaseFactory):
 
     Attributes:
         reader (rio_tiler.io.base.BaseReader): A rio-tiler reader. Defaults to `rio_tiler.io.Reader`.
-        path_dependency (Callable): Endpoint dependency defining `path` to pass to the reader init.
-        dataset_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining dataset overwriting options (e.g nodata).
-        layer_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining dataset indexes/bands/assets options.
-        render_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining image rendering options (e.g add_mask).
-        colormap_dependency (Callable): Endpoint dependency defining ColorMap options (e.g colormap_name).
-        process_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining image post-processing options (e.g rescaling, color-formula).
-        tms_dependency (Callable): Endpoint dependency defining TileMatrixSet to use.
         reader_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining BaseReader options.
-        environment_dependency (Callable): Endpoint dependency to define GDAL environment at runtime.
+        path_dependency (Callable): Endpoint dependency defining `path` to pass to the reader init.
+        layer_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining dataset indexes/bands/assets options.
+        dataset_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining dataset overwriting options (e.g nodata).
+        tile_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining tile options (e.g buffer, padding).
         stats_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining options for rio-tiler's statistics method.
         histogram_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining options for numpy's histogram method.
         img_preview_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining options for rio-tiler's preview method.
         img_part_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining options for rio-tiler's part/feature methods.
+        process_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining image post-processing options (e.g rescaling, color-formula).
+        rescale_dependency (Callable[..., Optional[RescaleType]]):
+        color_formula_dependency (Callable[..., Optional[str]]):
+        colormap_dependency (Callable): Endpoint dependency defining ColorMap options (e.g colormap_name).
+        render_dependency (titiler.core.dependencies.DefaultDependency): Endpoint dependency defining image rendering options (e.g add_mask).
+        environment_dependency (Callable): Endpoint dependency to define GDAL environment at runtime.
+        supported_tms (morecantile.defaults.TileMatrixSets): TileMatrixSets object holding the supported TileMatrixSets.
+        templates (Jinja2Templates): Jinja2 templates.
         add_preview (bool): add `/preview` endpoints. Defaults to True.
         add_part (bool): add `/bbox` and `/feature` endpoints. Defaults to True.
         add_viewer (bool): add `/map` endpoints. Defaults to True.
@@ -307,9 +312,12 @@ class TilerFactory(BaseFactory):
         self.bounds()
         self.info()
         self.statistics()
+        self.tilesets()
         self.tile()
-        self.tilejson()
+        if self.add_viewer:
+            self.map_viewer()
         self.wmts()
+        self.tilejson()
         self.point()
 
         # Optional Routes
@@ -318,9 +326,6 @@ class TilerFactory(BaseFactory):
 
         if self.add_part:
             self.part()
-
-        if self.add_viewer:
-            self.map_viewer()
 
     ############################################################################
     # /bounds
@@ -524,6 +529,229 @@ class TilerFactory(BaseFactory):
                         feature.properties.update({"statistics": stats})
 
             return fc.features[0] if isinstance(geojson, Feature) else fc
+
+    ############################################################################
+    # /tileset
+    ############################################################################
+    def tilesets(self):
+        """Register OGC tilesets endpoints."""
+
+        @self.router.get(
+            "/tiles",
+            response_model=TileSetList,
+            response_class=JSONResponse,
+            response_model_exclude_none=True,
+            responses={
+                200: {
+                    "content": {
+                        "application/json": {},
+                    }
+                }
+            },
+            summary="Retrieve a list of available raster tilesets for the specified dataset.",
+        )
+        async def tileset_list(
+            request: Request,
+            src_path=Depends(self.path_dependency),
+            reader_params=Depends(self.reader_dependency),
+            crs=Depends(CRSParams),
+            env=Depends(self.environment_dependency),
+        ):
+            """Retrieve a list of available raster tilesets for the specified dataset."""
+            with rasterio.Env(**env):
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
+                    bounds = src_dst.get_geographic_bounds(crs or WGS84_CRS)
+
+            collection_bbox = {
+                "lowerLeft": [bounds[0], bounds[1]],
+                "upperRight": [bounds[2], bounds[3]],
+                "crs": CRS_to_uri(crs or WGS84_CRS),
+            }
+
+            qs = [
+                (key, value)
+                for (key, value) in request.query_params._list
+                if key.lower() not in ["crs"]
+            ]
+            query_string = f"?{urlencode(qs)}" if qs else ""
+
+            tilesets = []
+            for tms in self.supported_tms.list():
+                tileset = {
+                    "title": f"tileset tiled using {tms} TileMatrixSet",
+                    "dataType": "map",
+                    "crs": self.supported_tms.get(tms).crs,
+                    "boundingBox": collection_bbox,
+                    "links": [
+                        {
+                            "href": self.url_for(
+                                request, "tileset", tileMatrixSetId=tms
+                            )
+                            + query_string,
+                            "rel": "self",
+                            "type": "application/json",
+                            "title": f"Tileset tiled using {tms} TileMatrixSet",
+                        },
+                        {
+                            "href": self.url_for(
+                                request,
+                                "tile",
+                                tileMatrixSetId=tms,
+                                z="{z}",
+                                x="{x}",
+                                y="{y}",
+                            )
+                            + query_string,
+                            "rel": "tile",
+                            "title": "Templated link for retrieving Raster tiles",
+                        },
+                    ],
+                }
+
+                try:
+                    tileset["links"].append(
+                        {
+                            "href": str(
+                                request.url_for("tilematrixset", tileMatrixSetId=tms)
+                            ),
+                            "rel": "http://www.opengis.net/def/rel/ogc/1.0/tiling-schemes",
+                            "type": "application/json",
+                            "title": f"Definition of '{tms}' tileMatrixSet",
+                        }
+                    )
+                except NoMatchFound:
+                    pass
+
+                tilesets.append(tileset)
+
+            data = TileSetList.model_validate({"tilesets": tilesets})
+            return data
+
+        @self.router.get(
+            "/tiles/{tileMatrixSetId}",
+            response_model=TileSet,
+            response_class=JSONResponse,
+            response_model_exclude_none=True,
+            responses={200: {"content": {"application/json": {}}}},
+            summary="Retrieve the raster tileset metadata for the specified dataset and tiling scheme (tile matrix set).",
+        )
+        async def tileset(
+            request: Request,
+            tileMatrixSetId: Annotated[
+                Literal[tuple(self.supported_tms.list())],
+                Path(
+                    description="Identifier selecting one of the TileMatrixSetId supported."
+                ),
+            ],
+            src_path=Depends(self.path_dependency),
+            reader_params=Depends(self.reader_dependency),
+            env=Depends(self.environment_dependency),
+        ):
+            """Retrieve the raster tileset metadata for the specified dataset and tiling scheme (tile matrix set)."""
+            tms = self.supported_tms.get(tileMatrixSetId)
+            with rasterio.Env(**env):
+                with self.reader(
+                    src_path, tms=tms, **reader_params.as_dict()
+                ) as src_dst:
+                    bounds = src_dst.get_geographic_bounds(tms.rasterio_geographic_crs)
+                    minzoom = src_dst.minzoom
+                    maxzoom = src_dst.maxzoom
+
+                    collection_bbox = {
+                        "lowerLeft": [bounds[0], bounds[1]],
+                        "upperRight": [bounds[2], bounds[3]],
+                        "crs": CRS_to_uri(tms.rasterio_geographic_crs),
+                    }
+
+                    tilematrix_limit = []
+                    for zoom in range(minzoom, maxzoom + 1, 1):
+                        matrix = tms.matrix(zoom)
+                        ulTile = tms.tile(bounds[0], bounds[3], int(matrix.id))
+                        lrTile = tms.tile(bounds[2], bounds[1], int(matrix.id))
+                        minx, maxx = (min(ulTile.x, lrTile.x), max(ulTile.x, lrTile.x))
+                        miny, maxy = (min(ulTile.y, lrTile.y), max(ulTile.y, lrTile.y))
+                        tilematrix_limit.append(
+                            {
+                                "tileMatrix": matrix.id,
+                                "minTileRow": max(miny, 0),
+                                "maxTileRow": min(maxy, matrix.matrixHeight),
+                                "minTileCol": max(minx, 0),
+                                "maxTileCol": min(maxx, matrix.matrixWidth),
+                            }
+                        )
+
+            qs = [(key, value) for (key, value) in request.query_params._list]
+            query_string = f"?{urlencode(qs)}" if qs else ""
+
+            links = [
+                {
+                    "href": self.url_for(
+                        request,
+                        "tileset",
+                        tileMatrixSetId=tileMatrixSetId,
+                    ),
+                    "rel": "self",
+                    "type": "application/json",
+                    "title": f"Tileset tiled using {tileMatrixSetId} TileMatrixSet",
+                },
+                {
+                    "href": self.url_for(
+                        request,
+                        "tile",
+                        tileMatrixSetId=tileMatrixSetId,
+                        z="{z}",
+                        x="{x}",
+                        y="{y}",
+                    )
+                    + query_string,
+                    "rel": "tile",
+                    "title": "Templated link for retrieving Raster tiles",
+                    "templated": True,
+                },
+            ]
+            try:
+                links.append(
+                    {
+                        "href": str(
+                            request.url_for(
+                                "tilematrixset", tileMatrixSetId=tileMatrixSetId
+                            )
+                        ),
+                        "rel": "http://www.opengis.net/def/rel/ogc/1.0/tiling-schemes",
+                        "type": "application/json",
+                        "title": f"Definition of '{tileMatrixSetId}' tileMatrixSet",
+                    }
+                )
+            except NoMatchFound:
+                pass
+
+            if self.add_viewer:
+                links.append(
+                    {
+                        "href": self.url_for(
+                            request,
+                            "map_viewer",
+                            tileMatrixSetId=tileMatrixSetId,
+                        )
+                        + query_string,
+                        "type": "text/html",
+                        "rel": "data",
+                        "title": f"Map viewer for '{tileMatrixSetId}' tileMatrixSet",
+                    }
+                )
+
+            data = TileSet.model_validate(
+                {
+                    "title": f"tileset tiled using {tileMatrixSetId} TileMatrixSet",
+                    "dataType": "map",
+                    "crs": tms.crs,
+                    "boundingBox": collection_bbox,
+                    "links": links,
+                    "tileMatrixSetLimits": tilematrix_limit,
+                }
+            )
+
+            return data
 
     ############################################################################
     # /tiles